Minor: Refactor
